Ccasin/user func spec modified vars fix (#350)
Changes how we calculate what system state is modified when using a user function spec.

The old story was that we analyzed the actual implementation of the function to see what state it touched, and treated everything it touched as both an input an an output to the function. This had a few issues. Most pressingly, it doesn't work at all for functions where the implementation is unavailable or can't be analyzed. It also meant we overestimated what state is changed by a function.

The new system is to look at the variables in the user-provided function spec. Uses of variables corresponding to the initial values of registers/memory mean those are inputs, and uses of variables corresponding to the final values of registers/memory mean those are updated by the function. This works well enough - the tests pass (I haven't added new tests since it is exercised by all our existing user function spec tests).