Fixes detection of intrinsic calls in BAP
We no longer rely on a BIL lifter being available, but instead look at
the subroutine in question since the information is there on whether it
is intrinsic or not.

The `Sub.intrinsic` tag doesn't seem to get populated anywhere, so for
now we can check that the subroutine's name is prefixed with
`intrinsic:`.